Code review feedback: bartoszmajsak - better structure
